% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_modules.R
\name{assignTrainingFolds}
\alias{assignTrainingFolds}
\title{Assigns each row in the input data a fold number randomly while ensuring the
same number of cases/controls are assigned to each fold.}
\usage{
assignTrainingFolds(data, classColname, nFolds, pipelineRunInformation)
}
\arguments{
\item{data}{A matrix/data.frame with a column named <class.colname>
corresponding to a binary indicator of case/control.}

\item{classColname}{The name of the column corresponding to a binary
indicator of case/control.}

\item{nFolds}{The number of folds to be assigned.}

\item{pipelineRunInformation}{The object originally created by
\code{\link{beginPipelineRun}}.}
}
\value{
A vector with elements corresponding to rows in data. Each element is
  a number between 1-\code{nFolds} and represents which cross-validation fold
  the row is assigned to.
}
\description{
Assigns each row in the input data a fold number randomly while ensuring the
same number of cases/controls are assigned to each fold.
}
