% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_modules.R
\name{beginPipelineRun}
\alias{beginPipelineRun}
\title{To be called at the beginning of a pipeline run.}
\usage{
beginPipelineRun(note, randomSeed = 42, log = TRUE, logFolderPath)
}
\arguments{
\item{note}{A string describing the pipeline run. This will be saved in the
pipeline run log.}

\item{randomSeed}{A number to be used as the random seed in pipeline function
calls.}

\item{log}{TRUE iff the pipeline log should be saved to file.}

\item{logFolderPath}{if log is TRUE, the log file will be saved in the folder
specified by logFolderPath.}
}
\value{
A list containing pipeline run information. Many of the pipeline
functions require this list to be passed in as an argument.
}
\description{
To be called at the beginning of a pipeline run.
}
