% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkMethodPackageInstalled}
\alias{checkMethodPackageInstalled}
\title{Title}
\usage{
checkMethodPackageInstalled(method, type)
}
\arguments{
\item{method}{}

\item{type}{}
}
\description{
Title
}
\keyword{internal}
