% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{fitBARTSurvival}
\alias{fitBARTSurvival}
\title{fitBARTSurvival}
\usage{
fitBARTSurvival(
  trainXs,
  trainTarget,
  testXs,
  pipelineRunInformation,
  sparse = FALSE,
  mcCores = 10,
  nTree = 100,
  k = 2,
  nSkip = 500,
  type = "pbart"
)
}
\arguments{
\item{trainXs}{A matrix/data.frame of X variables to be used in the model
fitting. Columns should correspond to variables.}

\item{trainTarget}{A matrix/data.frame including columns named 'Event' and
'time_to_event' corresponding to a binary indicator of event/censoring and
the time to event/censoring respectively.}

\item{testXs}{A matrix/data.frame of out-of-sample data to be used in model
evaluation. The set of variables should be the same as those in
\code{trainXs}.}

\item{pipelineRunInformation}{The object originally created by
\code{\link{beginPipelineRun}}.}

\item{sparse}{Set to \code{TRUE} if the sparsity inducing Dirichlet prior
should be used for splitting rule variable selection.}

\item{mcCores}{The number of cores to be used by the MCMC sampler.}

\item{nTree}{The number of trees to be used by BART.}

\item{k}{The value of the hyperparameter k.}

\item{nSkip}{The number of MCMC samples that should be treated as burn-in and
discarded.}

\item{type}{\code{'pbart'} (probit) or \code{'lbart'} (logit)}
}
\value{
The fit BART survival model object of type
  \code{\link[BART]{surv.bart}}.
}
\description{
fitBARTSurvival
}
