% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_modules.R
\name{fitCoxModel}
\alias{fitCoxModel}
\title{Fit glmnet penalised cox model on training data.}
\usage{
fitCoxModel(
  xs,
  target,
  penalty,
  weights,
  nFoldsCV = 3,
  foldID = FALSE,
  pipelineRunInformation,
  modelLabel,
  prescaled = FALSE,
  penaltyFactor = NULL
)
}
\arguments{
\item{xs}{A matrix/data.frame corresponding to X variables. Columns should
correspond to features.}

\item{target}{A matrix/data.frame with columns named 'Event' and
'time_to_event'. 'Event' should correspond to a binary label.
'time_to_event' should correspond to a positive numerical value
representing the survival time (when Event == 1) or the time-to-censoring
(when Event == 0)}

\item{penalty}{\code{'lasso'}/\code{'elnet'}/\code{'ridge'} A string
specifying the penalty.}

\item{weights}{A vector of weights corresponding to the training data rows to
be used in model fitting.}

\item{nFoldsCV}{The number of folds to be used in cross-validation (default
is 3)}

\item{foldID}{A vector of numbers specifying a cross-validation fold. Each
entry should correspond to a row in \code{xs}.}

\item{pipelineRunInformation}{The object originally created by
\code{\link{beginPipelineRun}}.}

\item{modelLabel}{A string that will act as a label for this function call.
The label will be used as a prefix for the saved object filenames in this
function call.}

\item{prescaled}{A boolean specifying if the columns in \code{xs} have been
scaled before the call to \code{\link{fitLogisticModel}} e.g. using
\code{scale()}. If set to TRUE, this sets \code{standardize = FALSE} in the
internal call to \code{cv.glmnet}}

\item{penaltyFactor}{A vector of numbers indicating variable-specific weights
for the lambda penalty.}
}
\value{
The fit penalised cox model glmnet object.
}
\description{
Fit glmnet penalised cox model on training data.
}
