% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_modules.R
\name{fitLogisticModel}
\alias{fitLogisticModel}
\title{This function is for an older version of MethylPipeR and is now deprecated.
It is currently still included to allow old scripts to run correctly.
Fits a penalised logistic regression model with cross-validation.}
\usage{
fitLogisticModel(
  xs,
  y,
  penalty,
  weights,
  nFoldsCV = 3,
  foldID = NULL,
  pipelineRunInformation,
  modelLabel,
  prescaled = FALSE,
  typeMeasure = "default",
  penaltyFactor = NULL
)
}
\arguments{
\item{xs}{A matrix/data.frame corresponding to X variables. Columns should
correspond to features.}

\item{y}{A matrix/data.frame with a column named 'Event' corresponding to a
binary label.}

\item{penalty}{\code{'lasso'}/\code{'elnet'}/\code{'ridge'} A string
specifying the penalty.}

\item{weights}{A vector of weights corresponding to the training data rows to
be used in model fitting.}

\item{nFoldsCV}{The number of folds to be used in cross-validation (default
is 3)}

\item{foldID}{A vector of numbers specifying a cross-validation fold. Each
entry should correspond to a row in \code{xs}.}

\item{pipelineRunInformation}{The object originally created by
\code{\link{beginPipelineRun}}}

\item{modelLabel}{A string that will act as a label for this function call.
The label will be used as a prefix in the 'penalty' and 'number of folds'
entries in pipeline.run.information. It will also be used as a prefix for
the saved object filenames in this function call.}

\item{prescaled}{A boolean specifying if the columns in \code{xs} have been
scaled before the call to fitLogisticModel e.g. using \code{scale()}. If
set to TRUE, this sets \code{standardize = FALSE} in the internal call to
\code{cv.glmnet}}

\item{typeMeasure}{A string specifying the argument to type.measure in
\code{cv.glmnet}. Defaults to \code{'default'}.}

\item{penaltyFactor}{A vector of numbers indicating variable-specific weights
for the lambda penalty.}
}
\value{
The glmnet logistic model object.
}
\description{
This function is for an older version of MethylPipeR and is now deprecated.
It is currently still included to allow old scripts to run correctly.
Fits a penalised logistic regression model with cross-validation.
}
