% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fitMPRModel}
\alias{fitMPRModel}
\title{Title}
\usage{
fitMPRModel(
  type,
  method,
  trainXs,
  trainY,
  testXs = NULL,
  testY = NULL,
  tteColname = "time_to_event",
  eventColname = "Event",
  parallel = FALSE,
  seed = NULL,
  ...
)
}
\arguments{
\item{type}{}

\item{method}{}

\item{trainXs}{}

\item{trainY}{}

\item{testXs}{}

\item{testY}{}

\item{parallel}{}

\item{seed}{}

\item{...}{}
}
\value{

}
\description{
Title
}
