% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fitMPRModelCV}
\alias{fitMPRModelCV}
\title{fitMPRModelCV}
\usage{
fitMPRModelCV(
  type,
  method,
  trainXs,
  trainY,
  testXs = NULL,
  testY = NULL,
  tteColname = "time_to_event",
  eventColname = "Event",
  parallel = FALSE,
  seed = NULL,
  save = TRUE,
  nFolds = 3,
  foldID = NULL,
  saveSuffix = NULL,
  ...
)
}
\arguments{
\item{type}{A string representing the type of model. Can be \code{'binary'},
\code{'survival'} or \code{'continuous'}.}

\item{method}{The modelling method. Currently supports \code{'glmnet'},
\code{'bart'} and \code{'rf'}.}

\item{trainXs}{The training data matrix/data.frame.}

\item{trainY}{The training response variable. This should be a vector if
\code{type = 'binary'}/\code{'continuous'} or a data.frame/matrix with
column names corresponding to \code{tteColname} and \code{eventColname} if
\code{type = 'survival'}.}

\item{testXs}{The test data matrix/data.frame.}

\item{testY}{The test response variable. This should be a vector if
\code{type = 'binary'}/\code{'continuous'} or a data.frame/matrix with
column names corresponding to \code{tteColname} and \code{eventColname} if
\code{type = 'survival'}.}

\item{tteColname}{A string corresponding to the time-to-event column name in
\code{trainY}/\code{testY}. Only required if \code{type == 'survival'}.}

\item{eventColname}{A string corresponding to the event column name in
\code{trainY}/\code{testY}. Only required if \code{type == 'survival'}.}

\item{parallel}{A boolean specifying whether parallel computation should be
used in model fitting.}

\item{seed}{An integer to set the random seed to for model fitting.}

\item{save}{A boolean specifying if the model object should be saved to the
logs.}

\item{nFolds}{The number of cross-validation folds to use.}

\item{foldID}{A vector of integers with length equal to the number of rows in
\code{trainXs}. Each element represents the fold number assigned to the
corresponding row.}

\item{saveSuffix}{Optional - a custom suffix to include in the file name of
the saved model object after the timestamp.}

\item{...}{The remaining parameters to be passed to the cross-validation 
function.}
}
\value{
Fits the specified model, selecting the best performing set of
hyperparameters from a k-fold cross-validation using each combination of
hyperparameters in a grid search.
}
\description{
fitMPRModelCV
}
