% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fitMPRModelIncremental}
\alias{fitMPRModelIncremental}
\title{Title}
\usage{
fitMPRModelIncremental(
  X,
  yColname,
  covColnames,
  scoreColname,
  family = "binomial"
)
}
\arguments{
\item{X}{}

\item{yColname}{}

\item{covColnames}{}

\item{scoreColname}{}

\item{family}{}
}
\value{

}
\description{
Title
}
