% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getBalancedCVFoldIDs}
\alias{getBalancedCVFoldIDs}
\title{For binary outcomes. Create a vector of (integer) fold assignments with equal numbers of cases (and controls) across folds.
Used to ensure there are as many cases as possible in all folds, particularly in imbalanced datasets.}
\usage{
getBalancedCVFoldIDs(data, classColname, nFolds, seed = NULL)
}
\arguments{
\item{data}{A dataframe containing a column specified by classColname}

\item{classColname}{A string corresponding to the binary class label (must contain only integers: 1 and 0)}

\item{nFolds}{The number of folds to assign}

\item{seed}{An integer to set for the random seed}
}
\value{
A vector of integers specifying the fold ID assigned to each row in data
}
\description{
For binary outcomes. Create a vector of (integer) fold assignments with equal numbers of cases (and controls) across folds.
Used to ensure there are as many cases as possible in all folds, particularly in imbalanced datasets.
}
