% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getGroupCVFoldIDs}
\alias{getGroupCVFoldIDs}
\title{Given a vector of group IDs 'groups', returns a vector where each element
is an integer representing the fold the corresponding element in 'groups'
is assigned to. Elements with the same group ID will always be assigned to
the same fold.}
\usage{
getGroupCVFoldIDs(groups, nFolds, seed = NULL)
}
\arguments{
\item{groups}{A vector of group IDs.}

\item{nFolds}{An integer representing the number of folds.}

\item{seed}{An integer to use as the random seed.}
}
\value{
A list containing foldIDs, groupToFoldMap and foldToNElementsMap.
}
\description{
Given a vector of group IDs 'groups', returns a vector where each element
is an integer representing the fold the corresponding element in 'groups'
is assigned to. Elements with the same group ID will always be assigned to
the same fold.
}
