% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_modules.R
\name{getNonZeroCoefficients}
\alias{getNonZeroCoefficients}
\title{This function is for an older version of MethylPipeR and is now deprecated.
It is currently still included to allow old scripts to run correctly.
Retrieve non-zero coefficients from a glmnet model}
\usage{
getNonZeroCoefficients(model, pipelineRunInformation, modelLabel)
}
\arguments{
\item{model}{A glmnet model object.}

\item{pipelineRunInformation}{The object originally created by beginPipelineRun.}

\item{modelLabel}{A string that will act as a label for this function call. The label will be used as a prefix for the saved object filenames in this function call.}
}
\value{
A data.frame containing the non-zero model coefficients.
}
\description{
This function is for an older version of MethylPipeR and is now deprecated.
It is currently still included to allow old scripts to run correctly.
Retrieve non-zero coefficients from a glmnet model
}
