% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{initLogs}
\alias{initLogs}
\title{To be called at the beginning of a MethylPipeR session. Initialises logging and model saving functionality.}
\usage{
initLogs(sessionLogFolder)
}
\arguments{
\item{sessionLogFolder}{The path of the folder that logs and model objects should be saved to. Must end in "/".}
}
\value{

}
\description{
To be called at the beginning of a MethylPipeR session. Initialises logging and model saving functionality.
}
