% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_modules.R
\name{loadData}
\alias{loadData}
\title{Load data from a file. Currently supports .rds and .csv files}
\usage{
loadData(dataFilepath)
}
\arguments{
\item{dataFilepath}{The filepath of the data to be loaded. Currently supports
'.rds' and '.csv' endings}
}
\value{
The loaded data object
}
\description{
Load data from a file. Currently supports .rds and .csv files
}
