% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logSessionLines}
\alias{logSessionLines}
\title{Writes lines to the session log.}
\usage{
logSessionLines(...)
}
\arguments{
\item{...}{The strings to be written to the session log. One string per line.}
}
\value{

}
\description{
Writes lines to the session log.
}
