% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{multipleROC}
\alias{multipleROC}
\title{multipleROC
Plot multiple ROC or PRROC curves with associated AUC values in the legend}
\usage{
multipleROC(
  labelsList,
  predictionsList,
  title,
  pipeline.run.information,
  yMetric = "tpr",
  xMetric = "fpr",
  auc = "ROC",
  legendPosition = "bottomright"
)
}
\arguments{
\item{labelsList}{todo}

\item{predictionsList}{todo}

\item{title}{todo}

\item{pipeline.run.information}{todo}

\item{yMetric}{todo}

\item{xMetric}{todo}

\item{auc}{todo}

\item{legendPosition}{todo}
}
\value{
todo
}
\description{
multipleROC
Plot multiple ROC or PRROC curves with associated AUC values in the legend
}
