% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_plotting.R
\name{plotMPRIncrementalModelConfusionMatrix}
\alias{plotMPRIncrementalModelConfusionMatrix}
\title{Plots confusion matrix metrics (true/false positives/negatives) at a range of
probability thresholds (from 0.1 to 1, inclusive, in steps of 0.1) given two 
sets of predicted probabilities (from two different models) and a binary 
target vector of ground truth class labels.}
\usage{
plotMPRIncrementalModelConfusionMatrix(
  nullModelResponse,
  fullModelResponse,
  target
)
}
\arguments{
\item{nullModelResponse}{A vector of predicted probabilities from the null 
model}

\item{fullModelResponse}{A vector of predicted probabilities from the full 
model}

\item{target}{A binary vector of ground truth class labels.}
}
\value{

}
\description{
Plots confusion matrix metrics (true/false positives/negatives) at a range of
probability thresholds (from 0.1 to 1, inclusive, in steps of 0.1) given two 
sets of predicted probabilities (from two different models) and a binary 
target vector of ground truth class labels.
}
