% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis_plot.R
\name{plotSingleVariableDensity}
\alias{plotSingleVariableDensity}
\title{plotSingleVariableDensity This old-format function is deprecated and will be
removed in a future version of MethylPipeR.}
\usage{
plotSingleVariableDensity(
  plotData,
  plotVariable,
  title,
  xLabel,
  yLabel,
  pipelineRunInformation
)
}
\arguments{
\item{plotData}{A dataframe containing the variable to be plotted}

\item{plotVariable}{A string specifying the variable to be plotted}

\item{title}{A string specifying the title for the plot}

\item{xLabel}{A string specifying the x-axis label}

\item{yLabel}{A string specifying the y-axis label}

\item{pipelineRunInformation}{a pipelineRunInformation object}
}
\value{
ggplot object
}
\description{
plotSingleVariableDensity This old-format function is deprecated and will be
removed in a future version of MethylPipeR.
}
