% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis_plot.R
\name{plotVariableByGroupDensity}
\alias{plotVariableByGroupDensity}
\title{plotVariableByGroupDensity}
\usage{
plotVariableByGroupDensity(
  plotData,
  plotVariable,
  groupVariable,
  title,
  xLabel,
  yLabel,
  pipelineRunInformation,
  filenamePrefix = NULL
)
}
\arguments{
\item{plotData}{data.frame containing variables to be plotted (at least two
column titled <plotVariable> and <groupVariable>)}

\item{plotVariable}{The name of the variable to be plotted}

\item{groupVariable}{The name of the variable representing the individual
groups to be plotted}

\item{title}{The title of the plot}

\item{xLabel}{The x-axis label}

\item{yLabel}{The y axis label}

\item{pipelineRunInformation}{the object originally returned by
\code{beginPipelineRun}}

\item{filenamePrefix}{The prefix for the file name of the saved plot object}
}
\value{
The ggplot object for the created grouped density plot
}
\description{
plotVariableByGroupDensity
}
