% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_modules.R
\name{predictCoxModel}
\alias{predictCoxModel}
\title{This function is for an older version of MethylPipeR and is now deprecated.
It is currently still included to allow old scripts to run correctly.
Calculate linear model predictions using glmnet Cox model.}
\usage{
predictCoxModel(model, xs, pipelineRunInformation, modelLabel)
}
\arguments{
\item{model}{A Cox glmnet model object.}

\item{xs}{A matrix/data.frame corresponding to X (variables). Columns should
correspond to features.}

\item{modelLabel}{A string that will act as a label for this function call.
The label will be used as a prefix for the saved object file names in this
function call.}
}
\value{
A vector of model predictions.
}
\description{
This function is for an older version of MethylPipeR and is now deprecated.
It is currently still included to allow old scripts to run correctly.
Calculate linear model predictions using glmnet Cox model.
}
