% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prediction.R
\name{predictMPRModel}
\alias{predictMPRModel}
\title{predictMPRModel}
\usage{
predictMPRModel(model, data, ...)
}
\arguments{
\item{model}{An MPRModel object. This is typically returned from a call to fitMPRModel.}

\item{data}{The data.frame/matrix that the model will be applied to.}

\item{...}{Other arguments to be passed to the method-specific prediction function.}
}
\value{
The result from the method-specific predict function.
}
\description{
predictMPRModel
}
