% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{saveMPRModelObject}
\alias{saveMPRModelObject}
\title{Saves a MPRModel object to the logging folder.}
\usage{
saveMPRModelObject(model, suffix = NULL)
}
\arguments{
\item{model}{The MPRModel object to be saved.}

\item{suffix}{Optional - a custom suffix to include in the filename after the timestamp.}
}
\value{

}
\description{
Saves a MPRModel object to the logging folder.
}
