% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fitMPRModelCV}
\alias{fitMPRModelCV}
\title{Title}
\usage{
fitMPRModelCV(
  type,
  method,
  trainXs,
  trainY,
  testXs = NULL,
  testY = NULL,
  tteColname = "time_to_event",
  eventColname = "Event",
  parallel = FALSE,
  seed = NULL,
  nFolds = 3,
  foldID = NULL,
  ...
)
}
\arguments{
\item{type}{}

\item{method}{}

\item{trainXs}{}

\item{trainY}{}

\item{testXs}{}

\item{testY}{}

\item{tteColname}{}

\item{eventColname}{}

\item{parallel}{}

\item{seed}{}

\item{nFolds}{}

\item{foldID}{}
}
\value{

}
\description{
Title
}
