% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fitMPRModelIncremental}
\alias{fitMPRModelIncremental}
\title{fitMPRModelIncremental}
\usage{
fitMPRModelIncremental(
  X,
  yColname,
  covColnames,
  scoreColname,
  family = "binomial"
)
}
\arguments{
\item{X}{A data.frame containing columns corresponding to \code{yColname},
\code{covColnames} and \code{scoreColname}.}

\item{yColname}{The name of the column representing the response variable.}

\item{covColnames}{A vector of names representing the covariate columns.}

\item{scoreColname}{The name of the score column.}

\item{family}{The GLM family string. \code{'binomial'} for binary response.
\code{'gaussian'} for continuous response.}
}
\value{
A list containing results for the null (covariates only) model and
  full model (score and covariates). For each model, the formula, response
  and \code{\link[stats]{glm}} model object are provided.
}
\description{
fitMPRModelIncremental
}
