% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_plotting.R
\name{plotMPRIncrementalModelROC}
\alias{plotMPRIncrementalModelROC}
\title{Title}
\usage{
plotMPRIncrementalModelROC(incrementalModelResult)
}
\arguments{
\item{incrementalModelResult}{}
}
\description{
Title
}
