% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis_plot.R
\name{plotSingleVariableDensity}
\alias{plotSingleVariableDensity}
\title{plotSingleVariableDensity}
\usage{
plotSingleVariableDensity(
  plotData,
  plotVariable,
  title,
  xLabel,
  yLabel,
  pipelineRunInformation
)
}
\arguments{
\item{pipelineRunInformation}{}
}
\value{
ggplot object
}
\description{
plotSingleVariableDensity
}
