% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_modules.R
\name{predictLogisticModel}
\alias{predictLogisticModel}
\title{This function is for an older version of MethylPipeR and is now deprecated.
It is currently still included to allow old scripts to run correctly.
Outputs predictions given a logistic model, data and labels.}
\usage{
predictLogisticModel(
  model,
  xs,
  pipelineRunInformation,
  modelLabel,
  predictType = "class"
)
}
\arguments{
\item{model}{A glmnet logistic regression model.}

\item{xs}{A matrix/data.frame corresponding to X variables. Columns should
correspond to features.}

\item{pipelineRunInformation}{The object originally created by
\code{\link{beginPipelineRun}}}

\item{modelLabel}{A string that will act as a label for this function call.
The label will be used as a prefix for the saved object file names in this
function call.}

\item{predictType}{A string specifying the prediction type (\code{'class'} or
\code{'response'}).}
}
\value{
A vector of model predictions.
}
\description{
This function is for an older version of MethylPipeR and is now deprecated.
It is currently still included to allow old scripts to run correctly.
Outputs predictions given a logistic model, data and labels.
}
