% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{saveMPRModelObject}
\alias{saveMPRModelObject}
\title{Saves a MPRModel object to the logging folder.}
\usage{
saveMPRModelObject(model)
}
\arguments{
\item{model}{The MPRModel object to be saved.}
}
\value{

}
\description{
Saves a MPRModel object to the logging folder.
}
